+++ packages/fighting-design/divider/src/interface.d.ts
@@ -1,5 +1,11 @@
 export type { DividerProps } from './props'
 
+/**
+ * 不同位置
+ */
 export type DividerPosition = 'left' | 'center' | 'right'
 
+/**
+ * 不同类型
+ */
 export type DividerType = 'dashed' | 'dotted' | 'double' | 'solid'